Delete confirmation & priority editing
Added a confirmation interface when deleting a task, so people can't just accidentally delete one.
Alongside this, I added a way to edit the priority level of a task.

+++ Security/Bound/Task System/Task System.js
@@ -1,5 +1,5 @@
 function AddTask() {
-  let html = HtmlService.createHtmlOutputFromFile("Task Manager/Add Task");
+  let html = HtmlService.createHtmlOutputFromFile("Add Task");
   html.setWidth(600);
   html.setHeight(800);
   return SpreadsheetApp.getUi().showModalDialog(html, "Task Interface");
@@ -33,7 +33,7 @@ function SubmitTask(inputData) {
   }
 
   s.getRange(loc, 2, 1, 7).setValues([["Backlog", Utilities.formatDate(new Date(), "GMT", 'dd MMMM yyyy'), inputData.deadline, inputData.title, inputData.description, "", inputData.priority]]);
-  s.getRange("A1").setValue(endLoc);
+  inputData.deadline = new Date(inputData.deadline);
   SendDiscordMessage("New", inputData);
   return "Task Added";
 }
@@ -72,9 +72,18 @@ function TaskManager(e) {
       OpenFile("Postpone", row, col, title);
       return range.setValue(false);
     case 13:
-      const inputData = DeleteTask(row);
-      SendDiscordMessage("Delete", inputData);
-      HasTask();
+      OpenFile("Priority", row, col, title);
+      return range.setValue(false);
+    case 14:
+      const response = SpreadsheetApp.getUi().alert("Delete?", "Do you wish to proceed? This action cannot be undone.\n", SpreadsheetApp.getUi().ButtonSet.YES_NO);
+
+      if (response == SpreadsheetApp.getUi().Button.YES) {
+        const inputData = DeleteTask(row);
+        SendDiscordMessage("Delete", inputData);
+        HasTask();
+      } else {
+        return range.setValue(false);
+      }
       return;
     default:
       return;
@@ -85,7 +94,7 @@ function TaskManager(e) {
  * Open a modal Dialog for the task manager
  */
 function OpenFile(type, row, col, title) {
-  template = HtmlService.createTemplateFromFile("Task Manager/Manager");
+  template = HtmlService.createTemplateFromFile("Manager");
   template.row = row;
   template.col = col;
   template.title = title;
@@ -96,6 +105,7 @@ function OpenFile(type, row, col, title) {
 
 /**
  * Used when postponing a deadline
+ * @returns {String}
  */
 function ChangeDeadline(inputData) {
   if (!inputData) throw new Error("Do not run this function from the editor");
@@ -104,17 +114,23 @@ function ChangeDeadline(inputData) {
   const types = ["Backlog", "In Progress", "Completed"];
   if (!types.includes(s.getRange(inputData.row, 2).getValue())) return SpreadsheetApp.getUi().alert("This is not a task");
 
+  inputData.deadline = new Date(inputData.deadline);
   s.getRange(inputData.row, 4).setValue(inputData.deadline);
   s.getRange(inputData.row, inputData.col).setValue(false);
+  
   inputData["title"] = s.getRange(inputData.row, 5).getValue();
   inputData["description"] = s.getRange(inputData.row, 6).getValue();
-  SendDiscordMessage("Postpone", inputData)
+  SendDiscordMessage("Postpone", inputData);
   return "Task Postponed";
 }
 
+/**
+ * @returns {String}
+ */
 function ChangeAssignment(inputData) {
   if (!inputData) throw new Error("Do not run this function from the editor");
 
+  if (!inputData.gmail.includes("@") || !inputData.gmail.includes(".")) return "Not a valid Gmail";
   const targetData = GetUserData(inputData.gmail);
   if (!targetData.row) return "User not found";
 
@@ -134,7 +150,7 @@ function ChangeAssignment(inputData) {
   inputData["description"] = s.getRange(inputData.row, 6).getValue();
   inputData["deadline"] = s.getRange(inputData.row, 4).getValue();
   SendDiscordMessage(assigned, inputData, targetData);
-  HasTask();
+  HasTask(assigned === "Assigned" ? true : false);
 
   return `${targetData.name} ${assigned}`;
 }
@@ -186,7 +202,7 @@ function CycleStatus(row) {
   if (s.getRange(row, 5).getValue() == "") return SpreadsheetApp.getUi().alert("This task is empty");
 
   const types = ["Backlog", "In Progress", "Completed"];
-  const type = s.getRange(row, 2).getValue();
+  let type = s.getRange(row, 2).getValue();
   if (!types.includes(s.getRange(row, 2).getValue())) return SpreadsheetApp.getUi().alert("This is not a task");
   if (types[2].includes(s.getRange(row, 2).getValue())) return SpreadsheetApp.getUi().alert("You cannot change the status of a completed task");
 
@@ -222,7 +238,6 @@ function CycleStatus(row) {
 
   const data = s.getRange(row, 3, 1, 6).getValues();
   DeleteTask(row, 11);
-  SendDiscordMessage("Status", inputData);
   console.log(`Data: ${data}`)
 
   // If there are no more free slots in a type
@@ -240,6 +255,26 @@ function CycleStatus(row) {
 
   s.getRange(newLoc, 3, 1, 6).setValues(data);
   s.getRange(newLoc, 2).setValue(type);
+  SendDiscordMessage("Status", inputData);
+}
+
+/**
+ * Edit the priority level of a task
+ * @param {Object} inputData - Data inputted by the user (priority, row, col, title)
+ * @returns {String}
+ */
+function ChangePriority(inputData) {
+  if (!inputData) throw new Error("Do not run this function from the editor");
+  const s = getCollect(1504741049);
+  if (s.getRange(inputData.row, 5).getValue() == "") return SpreadsheetApp.getUi().alert("This task is empty");
+
+  const types = ["Backlog", "In Progress", "Completed"];
+  if (!types.includes(s.getRange(inputData.row, 2).getValue())) return SpreadsheetApp.getUi().alert("This is not a task");
+  if (types[2].includes(s.getRange(inputData.row, 2).getValue())) return SpreadsheetApp.getUi().alert("You cannot change the priority of a completed task");
+
+  s.getRange(inputData.row, 8).setValue(inputData.priority);
+  SendDiscordMessage("Priority", inputData);
+  return "Priority Edited";
 }
 
 /**
@@ -247,11 +282,12 @@ function CycleStatus(row) {
  * @param {String} query - Cell where the reference is located
  * @returns {void}
  */
-function HasTask() {
+function HasTask(newTask = false) {
   const sheet = getCollect(2063800821);
   const taskSheet = getCollect(1504741049);
   const total_rows = sheet.getMaxRows();
   const task_rows = GetLastTaskRow(taskSheet, "Backlog");
+  let total_assignees = [];
 
   for (let i = 6; i < total_rows; i++) {
     const rank = sheet.getRange(i, 4).getValue();
@@ -260,18 +296,22 @@ function HasTask() {
       continue;
     }
     if (!rank) continue;
-    const name = sheet.getRange(i, 5);
-    const member = name.getValue();
+    const member = sheet.getRange(i, 5).getValue();
     if (!member) {
       sheet.getRange(i, 15).setValue(false);
       continue;
     }
     for (let j = 10; j <= task_rows; j++) {
-      const assignees = sheet.getRange(j, 7).getValue();
+      const assignees = taskSheet.getRange(j, 7).getValue();
       if (!assignees) continue;
+      if (!newTask) total_assignees.push(assignees);
       if (assignees.includes(member)) sheet.getRange(i, 15).setValue(true);
     }
-    if (name.getValue() != true) sheet.getRange(i, 15).setValue(false);
+
+    if (!newTask) {
+      if (!total_assignees.includes(sheet.getRange(i, 15).getValue())) sheet.getRange(i, 15).setValue(false);
+      total_assignees = [];
+    }
   }  
 }
 
@@ -373,7 +413,6 @@ function RemoveAssignee(str, name) {
     const escapedName = name.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&');
     const regex = new RegExp(`\\s*${escapedName}\\s*,?|,?\\s*${escapedName}\\s*`, 'g');
     const newStr = str.replace(regex, '').trim();
-    console.log(`New Assignees: ${newStr}`);
     return newStr;
   } else {
     return str;
@@ -447,6 +486,7 @@ function GetFirstTaskRow(s, type) {
 
 function GetUserData(query) {
   if (!query) throw new Error("Do not run this function from the editor");
+  if (query == "N/A") throw new Error("You cannot target Site Management");
   const s = getCollect(2063800821);
   const r = s.getRange(6, 8, s.getMaxRows(), 1).getValues();
   let data = {};
